requirements.txt
APScheduler==3.6.3
backports.zoneinfo==0.2.1
base58==2.1.1
cachetools==4.2.2
certifi==2021.10.8
charset-normalizer==2.0.12
cytoolz==0.11.2
ecdsa==0.17.0
eth-hash==0.3.2
eth-keys==0.4.0
eth-typing==3.0.0
eth-utils==2.0.0
idna==3.3
more-itertools==8.12.0
numpy==1.22.3
pandas==1.4.2
py-bip39-bindings==0.1.9
py-ed25519-bindings==1.0.1
py-sr25519-bindings==0.1.4
pycryptodome==3.14.1
python-dateutil==2.8.2
python-telegram-bot==13.12
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
requests==2.27.1
scalecodec==1.0.35
six==1.16.0
substrate-interface==1.2.3
toolz==0.11.2
tornado==6.1
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.9
websocket-client==1.3.2
xxhash==2.0.2
